[workbox-precaching] Moved some file to sub-folder
