+++ types/workbox-precaching/precacheAndRoute.d.ts
@@ -1,6 +1,6 @@
 import { FetchListenerOptions } from "./types/FetchListenerOptions";
-import { PrecacheEntry } from "./PrecacheEntry";
+import { PrecacheEntry } from "./types/PrecacheEntry";
 
 export function precacheAndRoute(entries: Array<PrecacheEntry | string>, options?: FetchListenerOptions): void;
 
-export * from "./PrecacheEntry";
+export * from "./types/PrecacheEntry";
