+++ types/workbox-precaching/types/CleanupResult.d.ts
