+++ types/workbox-precaching/types/PrecacheEntry.d.ts
