+++ types/workbox-precaching/types/URLManipulation.d.ts
